Corrected calculation for match share and difference.

+++ src/features/statistics/models/PlayerMatchDayDataDto.cs
@@ -2,7 +2,7 @@ namespace skat_back.features.statistics.models;
 
 public record PlayerMatchDayDataDto(
     string Name,
-    int MatchShare,
+    float MatchShare,
     int TotalPoints,
     List<SeriesDto> Series
 );
